refactor: DNS module refactored into package
